RAG/requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.35.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.3.0
langchain-core>=0.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
accelerate>=0.24.0
bitsandbytes>=0.41.0
rank-bm25>=0.2.2
pandas>=1.5.0
numpy>=1.24.0 